Fix checkboxlist not showing value if options label and value is integer
Checkboxlist will not show value from `$options` if it's an integer. 
For example, if the `$fieldOptions` is an array with all integer values:
```php
[ 
2000 => 2000, 
2001 => 2001,
2009 => 2009
]
```
The `Checkboxlist` widget won't add anything to the `<label>` tag.

This is because there's a check here that make sure the $option is an array. (I'm not sure why though)
```php 
/* check is_string */
if (is_string($option)) $option = [$option];
/* output to label */
 <label for="<?= $checkboxId ?>">
        <?= e(trans($option[0])) ?>
  </label>
```

If the label & value is not a string, then accessing`$option[0]` will not work.

Adding `is_int()` check, fixed the problem.
`if (is_string($option) || is_int($option))`

My workaround was to cast all the `getSomethingOptions` results into string. 
But since the fix is pretty small I thought I give it a try to propose this PR.